in memoriam desktop css done
