+++ frontend/package.json
@@ -11,7 +11,7 @@
     "@emotion/styled": "^11.10.6",
     "@mui/icons-material": "^5.11.16",
     "@mui/material": "^5.12.1",
-    "axios": "^1.3.5",
+    "axios": "^1.4.0",
     "dayjs": "^1.11.7",
     "leaflet": "^1.9.3",
     "prop-types": "^15.8.1",
