--- frontend/src/components/InMemoriam.module.css
@@ -1,159 +1,0 @@
-.card {
-  width: 70vw;
-  height: 36vw;
-  border-radius: 20px;
-  flex-direction: column;
-  background-color: rgba(0, 0, 0, 0.535);
-}
-
-.containerGauche {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-
-.containerDroite {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  width: 100%;
-  justify-content: center;
-  align-content: center;
-}
-
-.button {
-  font-size: 50px;
-  color: rgba(190, 180, 180, 0.452);
-  cursor: pointer;
-}
-
-.linkWiki {
-  color: white;
-  text-decoration: none;
-}
-
-.linkWiki:hover {
-  cursor: pointer;
-  color: #9364d4;
-  transition: color 0.2s ease-out;
-}
-
-.profilePhoto {
-  width: 100px;
-  height: 100px;
-  border-radius: 100%;
-  margin-top: 3px;
-}
-
-.imageShip {
-  height: 450px;
-  width: 625px;
-  border-radius: 16px;
-  margin-left: 15px;
-  margin-top: 4px;
-}
-
-.astronautes {
-  margin-left: 650px;
-  display: inline-flex;
-  justify-content: space-evenly;
-}
-
-.title {
-  margin-top: 6%;
-}
-
-.date {
-  display: flex;
-  margin-top: 3px;
-  margin-bottom: 0;
-  font-size: 50px;
-  background-color: rgba(0, 0, 0, 0.404);
-  border-radius: 20px;
-  width: 500px;
-  text-align: center;
-  justify-content: center;
-}
-
-.badgeMission {
-  width: 150px;
-  height: 140px;
-  margin-top: 8px;
-  /*   margin-left: 200px;
-  margin-top: 10px; */
-}
-
-.nameOfShip {
-  margin-top: 40px;
-
-  font-size: 40px;
-  background-color: rgba(0, 0, 0, 0.404);
-  border-radius: 20px;
-  width: 500px;
-  height: 60px;
-}
-
-.astroCard {
-  width: 196px;
-  height: 180px;
-  background-color: rgba(0, 0, 0, 0.404);
-  border-radius: 20px;
-  margin-bottom: 20px;
-  margin: 1% 3% 0% 1%;
-}
-
-.astronautContainer {
-  display: inline-flex;
-}
-
-.astroName {
-  margin-top: initial;
-}
-
-.inMemoContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.flexmess {
-  margin-top: -3%;
-  justify-content: center;
-}
-
-/* .rightButtonContainer {
-  margin-bottom: 1000px;
-} */
-
-.CONTENAIRCARDBTN {
-  display: flex;
-  align-items: center;
-}
-
-@media (max-width: 426px) {
-  .citation {
-    display: none;
-  }
-
-  .card {
-    width: 816px;
-    height: 1396px;
-    margin-top: -3%;
-  }
-
-  .containerDroite {
-    margin-top: 85%;
-    margin-left: -88%;
-  }
-
-  .containerGauche {
-    margin-left: 10%;
-    margin-top: 3%;
-  }
-
-  .title {
-    margin-top: 68%;
-    margin-right: -93%;
-    font-size: 49px;
-  }
-}
