+++ frontend/src/components/mapAgencies/MapAgencies.css
@@ -1,16 +1,8 @@
 @import url("https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;300;400;500&display=swap");
 
-.leaflet-container {
-  height: 100%;
-  width: 95%;
-
-  border-radius: 15px;
-  margin-bottom: 20px;
-}
-
 .mapContainer {
   width: 60vw;
-  height: 67.5vh;
+  height: 100%;
   border-radius: 15px;
   background-color: rgba(0, 0, 0, 0.38);
   display: flex;
@@ -19,7 +11,14 @@
   justify-content: center;
   font-family: "Poppins", sans-serif;
   font-size: 15px;
-  color: white;
+  color: whitesmoke;
+}
+
+.leaflet-container {
+  height: 100%;
+  width: 95%;
+  border-radius: 15px;
+  margin-bottom: 20px;
 }
 
 .agencieTitle {
@@ -40,7 +39,7 @@
   padding-bottom: 8px;
   font-size: 20px;
   font-weight: 100;
-  color: white;
+  color: whitesmoke;
 }
 
 .leaflet-popup-content-wrapper {
