+++ frontend/src/components/spacecraft-container/SpaceCraftContainer.module.css
@@ -11,13 +11,13 @@
   display: flex;
   align-items: center;
   justify-content: space-between;
-  height: 100%;
   align-items: stretch;
 }
 
 .leftContainer,
 .rightContainer {
   width: 45%;
+  height: 100%;
 }
 
 .inuseInfo {
