+++ frontend/src/components/spacecraft-container/spacecraft-card/SpacecraftCard.module.css
@@ -9,7 +9,7 @@
 .imgContainer {
   border-radius: 10px;
 
-  height: 25vh;
+  height: 30vh;
   overflow: hidden;
   display: flex;
   border-radius: 10px;
