+++ frontend/src/pages/InMemoriamPage.jsx
@@ -1,11 +1,19 @@
 import React from "react";
-import InMemorial from "../components/InMemoriam";
+import InMemorial from "../components/inmemoriam/InMemoriam";
 import styles from "./InMemoriamPage.module.css";
+import Article from "../components/article/Article";
 
 export default function InMemoriamPage() {
   return (
     <div className={styles.memoriamPage}>
-      <InMemorial />
+      <div className={styles.inmemoriamAndArticlesContainer}>
+        <div className={styles.inmemoriamContainer}>
+          <InMemorial />
+        </div>
+        <div className={styles.articlesContainer}>
+          <Article />
+        </div>
+      </div>
     </div>
   );
 }
