+++ frontend/src/pages/InMemoriamPage.module.css
@@ -5,7 +5,29 @@
   background-size: cover;
   width: 100vw;
   height: 100vh;
-  color: white;
-  display: flex;
+  color: whitesmoke;
   font-family: "Poppins", sans-serif;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.inmemoriamAndArticlesContainer {
+  display: flex;
+  flex-direction: row;
+  align-items: stretch;
+  justify-content: space-between;
+  gap: 5vw;
+  width: 90vw;
+  max-height: 600px;
+  margin: 0;
+  margin-top: 150px;
+}
+
+.inmemoriamContainer {
+  width: 70%;
+}
+
+.articlesContainer {
+  width: 30%;
 }
