+++ frontend/src/pages/SpaceCraft.module.css
@@ -17,7 +17,7 @@
   justify-content: space-between;
   gap: 5vw;
   width: 90vw;
-  max-height: 600px;
+  height: 65vh;
   font-family: "Poppins", sans-serif;
   color: white;
   text-align: center;
